feat: Update todo component to display item descriptions
The code changes modify the todo component's HTML template to display the description of each item instead of just the item itself. This change improves the user experience by providing more context for each todo item.

Based on the recent user commits and repository commits, the commit message follows the established convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or enhancement.

```