feat: Add delete functionality to todo component
The code changes in this commit add the ability to delete a todo list item in the todo component. The `deleteList` method is implemented in the `ApiService` to send a DELETE request to the server, and the corresponding HTML template is updated to include a "DELETE USER" button.

Based on the recent user commits and repository commits, the commit message follows the established convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or enhancement.