feat: Add lazy loading for images in todo component
The code changes in this commit modify the HTML template of the todo component to add lazy loading for images. This improves the page load performance by loading images only when they are visible to the user.

Based on the recent user commits and repository commits, the commit message follows the established convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or enhancement.